fix(stylelint-plugin): address PR feedback [KHCP-10186]
