+++ .github/workflows/test.yaml
@@ -45,10 +45,8 @@ jobs:
       - name: Lint
         run: yarn lint
 
-      - name: Unit test stylelint-plugin
-        run: |
-          yarn --cwd stylelint-plugin install --frozen-lockfile
-          yarn --cwd stylelint-plugin test:unit
+      - name: Run unit tests
+        run: yarn test:unit:coverage
 
       - name: Build style dictionary
         run: yarn build
